requirements.txt
numpy
pymysql
scikit-surprise
